model: Split out classes contained in `ResolvedLicenseInfo` to separate files
This helps to get a better overview over the many classes.
