+++ model/src/main/kotlin/licenses/ResolvedLicenseInfo.kt
@@ -23,12 +23,9 @@ import org.ossreviewtoolkit.model.CopyrightFinding
 import org.ossreviewtoolkit.model.Identifier
 import org.ossreviewtoolkit.model.LicenseSource
 import org.ossreviewtoolkit.model.Provenance
-import org.ossreviewtoolkit.model.TextLocation
 import org.ossreviewtoolkit.model.config.CopyrightGarbage
-import org.ossreviewtoolkit.model.config.LicenseFindingCuration
 import org.ossreviewtoolkit.model.config.PathExclude
 import org.ossreviewtoolkit.utils.ort.CopyrightStatementsProcessor
-import org.ossreviewtoolkit.utils.ort.DeclaredLicenseProcessor
 import org.ossreviewtoolkit.utils.spdx.SpdxExpression
 import org.ossreviewtoolkit.utils.spdx.SpdxSingleLicenseExpression
 import org.ossreviewtoolkit.utils.spdx.model.SpdxLicenseChoice
@@ -138,194 +135,6 @@ data class ResolvedLicenseInfo(
 }
 
 /**
- * Resolved information for a single license.
- */
-data class ResolvedLicense(
-    /**
-     * The license.
-     */
-    val license: SpdxSingleLicenseExpression,
-
-    /**
-     * The set of original declared license that were [processed][DeclaredLicenseProcessor] to this [license], or an
-     * empty list, if this [license] was not modified during processing.
-     */
-    val originalDeclaredLicenses: Set<String>,
-
-    /**
-     * The original SPDX expressions of this license.
-     */
-    val originalExpressions: Set<ResolvedOriginalExpression>,
-
-    /**
-     * All text locations where this license was found.
-     */
-    val locations: Set<ResolvedLicenseLocation>
-) {
-    /**
-     * The sources where this license was found.
-     */
-    val sources: Set<LicenseSource> = originalExpressions.mapTo(mutableSetOf()) { it.source }
-
-    /**
-     * True, if this license was [detected][LicenseSource.DETECTED] and all [locations] have matching path excludes.
-     */
-    val isDetectedExcluded by lazy {
-        LicenseSource.DETECTED in sources && locations.all { it.matchingPathExcludes.isNotEmpty() }
-    }
-
-    init {
-        require(sources.isNotEmpty()) {
-            "A resolved license must have at least one license source."
-        }
-    }
-
-    /**
-     * Return all resolved copyrights for this license. Copyright findings that are excluded by [PathExclude]s are
-     * [omitted][omitExcluded] by default. The copyrights are [processed][process] by default using the
-     * [CopyrightStatementsProcessor].
-     */
-    fun getResolvedCopyrights(process: Boolean = true, omitExcluded: Boolean = true): List<ResolvedCopyright> {
-        val resolvedCopyrightFindings = locations.flatMap { location ->
-            location.copyrights.filter { copyright ->
-                !omitExcluded || copyright.matchingPathExcludes.isEmpty()
-            }
-        }
-
-        return resolvedCopyrightFindings.toResolvedCopyrights(process)
-    }
-
-    /**
-     * Return all copyright statements associated to this license. Copyright findings that are excluded by
-     * [PathExclude]s are [omitted][omitExcluded] by default. The copyrights are [processed][process] by default
-     * using the [CopyrightStatementsProcessor].
-     */
-    @JvmOverloads
-    fun getCopyrights(process: Boolean = true, omitExcluded: Boolean = true): Set<String> =
-        getResolvedCopyrights(process, omitExcluded).mapTo(mutableSetOf()) { it.statement }
-
-    /**
-     * Filter all excluded copyrights. Copyrights which have
-     * [matching path excludes][ResolvedCopyrightFinding.matchingPathExcludes] are removed.
-     */
-    fun filterExcludedCopyrights(): ResolvedLicense =
-        copy(
-            locations = locations.mapTo(mutableSetOf()) { location ->
-                location.copy(
-                    copyrights = location.copyrights.filterTo(mutableSetOf()) { it.matchingPathExcludes.isEmpty() }
-                )
-            }
-        )
-
-    /**
-     * Filter all excluded original expressions. Detected license findings which have
-     * [matching path excludes][ResolvedCopyrightFinding.matchingPathExcludes] are removed. If the resolved license
-     * becomes empty, then null is returned.
-     */
-    fun filterExcludedOriginalExpressions(): ResolvedLicense? {
-        if (LicenseSource.DETECTED !in sources) return this
-
-        val filteredOriginalExpressions = originalExpressions.filterNotTo(mutableSetOf()) { it.isDetectedExcluded }
-        if (filteredOriginalExpressions.isEmpty()) return null
-
-        return copy(originalExpressions = filteredOriginalExpressions)
-    }
-}
-
-/**
- * A resolved text location.
- */
-data class ResolvedLicenseLocation(
-    /**
-     * The provenance of the file.
-     */
-    val provenance: Provenance,
-
-    /**
-     * The text location relative to the root of the VCS or source artifact defined by [provenance].
-     */
-    val location: TextLocation,
-
-    /**
-     * The applied [LicenseFindingCuration], or null if none were applied.
-     */
-    val appliedCuration: LicenseFindingCuration?,
-
-    /**
-     * All [PathExclude]s matching this [location].
-     */
-    val matchingPathExcludes: List<PathExclude>,
-
-    /**
-     * All copyright findings associated to this license location, excluding copyright garbage.
-     */
-    val copyrights: Set<ResolvedCopyrightFinding>
-)
-
-/**
- * A resolved copyright.
- */
-data class ResolvedCopyright(
-    /**
-     * The resolved copyright statement.
-     */
-    val statement: String,
-
-    /**
-     * The resolved findings for this copyright. The statements in the findings can be different to [statement] if they
-     * were processed by the [CopyrightStatementsProcessor].
-     */
-    val findings: Set<ResolvedCopyrightFinding>
-)
-
-/**
- * A resolved copyright finding.
- */
-data class ResolvedCopyrightFinding(
-    /**
-     * The copyright statement.
-     */
-    val statement: String,
-
-    /**
-     * The location where this copyright was found.
-     */
-    val location: TextLocation,
-
-    /**
-     * All [PathExclude]s matching this [location].
-     */
-    val matchingPathExcludes: List<PathExclude>
-)
-
-/**
- * A resolved original expression.
- */
-data class ResolvedOriginalExpression(
-    /**
-     * The license expression.
-     */
-    val expression: SpdxExpression,
-
-    /**
-     * The license source.
-     */
-    val source: LicenseSource,
-
-    /**
-     * Indicate whether all license findings corresponding to [expression] are excluded. Must be false if [source] does
-     * not equal [LicenseSource.DETECTED].
-     */
-    val isDetectedExcluded: Boolean = false
-) {
-    init {
-        require(!isDetectedExcluded || source == LicenseSource.DETECTED) {
-            "For license source '$source' the flag isDetectedExcluded must not be set."
-        }
-    }
-}
-
-/**
  * Filter all excluded licenses and copyrights. Licenses are removed if they are only
  * [detected][LicenseSource.DETECTED] and all [locations][ResolvedLicense.locations] have
  * [matching path excludes][ResolvedLicenseLocation.matchingPathExcludes]. Copyrights are removed if all
@@ -338,21 +147,3 @@ fun List<ResolvedLicense>.filterExcluded() =
         resolvedLicense.sources != setOf(LicenseSource.DETECTED) ||
                 resolvedLicense.locations.any { it.matchingPathExcludes.isEmpty() }
     }.map { it.filterExcludedCopyrights() }
-
-private fun Collection<ResolvedCopyrightFinding>.toResolvedCopyrights(process: Boolean): List<ResolvedCopyright> {
-    val allStatements = map { it.statement }
-    val processedStatements = if (process) {
-        CopyrightStatementsProcessor().process(allStatements).toMap()
-    } else {
-        allStatements.associateBy({ it }, { mutableSetOf(it) })
-    }
-
-    return processedStatements.mapValues { (_, originalStatements) ->
-        filter { it.statement in originalStatements }
-    }.filterValues { it.isNotEmpty() }.entries.map { (statement, findings) ->
-        ResolvedCopyright(statement, findings.toSet())
-    }
-}
-
-private fun CopyrightStatementsProcessor.Result.toMap(): Map<String, Set<String>> =
-    processedStatements + unprocessedStatements.associateWith { setOf(it) }
